# txn_test_gen_plugin thread pool