--- a/plugins/txn_test_gen_plugin/txn_test_gen_plugin.cpp
+++ b/plugins/txn_test_gen_plugin/txn_test_gen_plugin.cpp
@@ -93,21 +93,16 @@ struct txn_test_gen_plugin_impl {
    uint64_t _total_us = 0;
    uint64_t _txcount = 0;
 
-   int _remain = 0;
-
    std::shared_ptr<boost::asio::io_context>             gen_ioc;
    optional<io_work_t>                                  gen_ioc_work;
    uint16_t                                             thread_pool_size;
    optional<boost::asio::thread_pool>                   thread_pool;
    std::shared_ptr<boost::asio::high_resolution_timer>  timer;
 
-   void push_next_transaction(const std::shared_ptr<std::vector<signed_transaction>>& trxs, size_t index, const std::function<void(const fc::exception_ptr&)>& next ) {
+   void push_next_transaction(const std::shared_ptr<std::vector<signed_transaction>>& trxs, const std::function<void(const fc::exception_ptr&)>& next ) {
       chain_plugin& cp = app().get_plugin<chain_plugin>();
 
-      const int overlap = 20;
-      size_t end = std::min(index + overlap, trxs->size());
-      _remain = end - index;
-      for (int i = index; i < end; ++i) {
+      for (int i = 0; i < trxs->size(); ++i) {
          cp.accept_transaction( packed_transaction(trxs->at(i)), [=](const fc::static_variant<fc::exception_ptr, transaction_trace_ptr>& result){
             if (result.contains<fc::exception_ptr>()) {
                next(result.get<fc::exception_ptr>());
@@ -115,14 +110,6 @@ struct txn_test_gen_plugin_impl {
                if (result.contains<transaction_trace_ptr>() && result.get<transaction_trace_ptr>()->receipt) {
                   _total_us += result.get<transaction_trace_ptr>()->receipt->cpu_usage_us;
                   ++_txcount;
-               }
-               --_remain;
-               if (_remain == 0 ) {
-                  if (end < trxs->size()) {
-                     push_next_transaction(trxs, index + overlap, next);
-                  } else {
-                     next(nullptr);
-                  }
                }
             }
          });
@@ -132,7 +119,7 @@ struct txn_test_gen_plugin_impl {
    void push_transactions( std::vector<signed_transaction>&& trxs, const std::function<void(fc::exception_ptr)>& next ) {
       auto trxs_copy = std::make_shared<std::decay_t<decltype(trxs)>>(std::move(trxs));
       app().post(priority::low, [this, trxs_copy, next]() {
-         push_next_transaction(trxs_copy, 0, next);
+         push_next_transaction(trxs_copy, next);
       });
    }